+++ src/python/WorkflowMgmt/WorkflowSchemaHandlers/StepChainWfSchemaHandler.py
@@ -2,26 +2,32 @@ import re
 from logging import Logger
 from collections import defaultdict
 
+from Utilities.IteratorTools import filterKeys, mapValues, sortByKeys
+from WorkflowMgmt.WorkflowSchemaHandlers.BaseWfSchemaHandler import BaseWfSchemaHandler
+
 from typing import Optional, Union, Any, Tuple, Callable
 
-from Utilities.IteratorTools import filterKeys, mapValues, sortByKeys
-from Components.Workload.BaseWorkloadHandler import BaseWorkloadHandler
 
-
-class BaseChainWorkloadHandler(BaseWorkloadHandler):
+class StepChainWfSchemaHandler(BaseWfSchemaHandler):
     """
-    __BaseChainWorkloadHandler__
-    General Abstract Base Class for building the concrete workload handlers based on the chain request type
+    __StepChainWfSchemaHandler__
+    General API for handling the request data of step chain request type
     """
 
     def __init__(self, wfSchema: dict, logger: Optional[Logger] = None) -> None:
         try:
-            super().__init__(wfSchema, logger=logger)
+            super().__init__(wfSchema, logger)
+            self.includeHEPCloudInSiteWhiteList = True
+
             self.base = self.wfSchema["RequestType"].replace("Chain", "")
             self.chainKeys = sorted(filter(re.compile(f"^{self.base}\d+$").search, self.wfSchema))
 
+            self.logMsg = {
+                "reduceLargeOutput": "The output size of task is expected to be too large : %d x %.2f kB = %.2f GB > %f GB. Reducing to %d"
+            }
+
         except Exception as error:
-            raise Exception(f"Error initializing {self.__class__.__name__}\n{str(error)}")
+            raise Exception(f"Error initializing StepChainWfSchemaHandler\n{str(error)}")
 
     def getChainValues(self, key: str, f: Optional[Callable] = None) -> dict:
         """
@@ -262,4 +268,60 @@ class BaseChainWorkloadHandler(BaseWorkloadHandler):
 
         except Exception as error:
             self.logger.error("Failed to get the computing time")
+            self.logger.error(str(error))
+
+    def checkSplitting(self, splittings: dict) -> Tuple[bool, list]:
+        """
+        The function to check the splittings sizes and if any action is required
+        :param splittings: splittings schema
+        :return: if to hold and a list of modified splittings
+        """
+        try:
+            modifiedSplittings = []
+
+            nCores = self.getMulticore()
+            totalGbSpaceLimit = nCores * self.unifiedConfiguration.get("GB_space_limit")
+
+            sizePerEvent = self.wfSchema.get("SizePerEvent", 0) / (1024.0 ** 2.0)
+            for splitting in splittings:
+                params = splitting.get("splitParams", {})
+
+                eventsPerJob = params.get("events_per_job", 0)
+                size = eventsPerJob * sizePerEvent
+
+                if size > totalGbSpaceLimit:
+                    maxTaskEventsPerJob = int(totalGbSpaceLimit / sizePerEvent)
+                    splitting["splitParams"]["events_per_job"] = maxTaskEventsPerJob
+                    modifiedSplittings.append(splitting)
+
+                    self.logger.info(
+                        self.logMsg["reduceLargeOutput"],
+                        eventsPerJob,
+                        sizePerEvent,
+                        size,
+                        totalGbSpaceLimit,
+                        maxTaskEventsPerJob,
+                    )
+
+            return False, modifiedSplittings
+
+        except Exception as error:
+            self.logger.error("Failed to check splittings")
+            self.logger.error(str(error))
+
+    def writeDatasetPatternName(self, elements: list) -> str:
+        """
+        The function to write the dataset pattern name from given elements
+        :param elements: dataset name elements — name, acquisition era, processing string, version, tier
+        :return: dataset name
+        """
+        try:
+            if (elements[3] == "v*" and all(element == "*" for element in elements[1:3])) or (
+                elements[3] != "v*" and elements[2] == "*"
+            ):
+                return None
+            return f"/{elements[0]}/{'-'.join(elements[1:4])}/{elements[4]}"
+
+        except Exception as error:
+            self.logger.error("Failed to write dataset pattern name for %s", elements)
             self.logger.error(str(error))
